Refactors for separate tag module and saves tags with relative top and left

+++ db/migrate/20160531221420_create_tags.rb
@@ -1,8 +1,8 @@
 class CreateTags < ActiveRecord::Migration
   def change
     create_table :tags do |t|
-      t.integer :top, null: false
-      t.integer :left, null: false
+      t.float :top, null: false
+      t.float :left, null: false
       t.string :character, null: false
 
       t.timestamps null: false
